+++ src/WebJobs.Extensions.DurableTask/LocalHttpListener.cs
@@ -22,9 +22,18 @@ namespace Microsoft.Azure.WebJobs.Extensions.DurableTask
     {
         private const int DefaultPort = 17071;
 
+        // Pick a large, fixed range of ports that are going to be valid in all environment.
+        // Avoiding ports below 1024 as those are blocked by app service sandbox.
+        // Ephemeral ports for most OS start well above 32768. See https://www.ncftp.com/ncftpd/doc/misc/ephemeral_ports.html
+        private const int MinPort = 30000;
+        private const int MaxPort = 31000;
+
         private readonly Func<HttpRequestMessage, Task<HttpResponseMessage>> handler;
         private readonly EndToEndTraceHelper traceHelper;
         private readonly DurableTaskOptions durableTaskOptions;
+        private readonly Random portGenerator;
+        private readonly HashSet<int> attemptedPorts;
+
         private IWebHost localWebHost;
 
         public LocalHttpListener(
@@ -39,6 +48,8 @@ namespace Microsoft.Azure.WebJobs.Extensions.DurableTask
             // Set to a non null value
             this.InternalRpcUri = new Uri($"http://uninitialized");
             this.localWebHost = new NoOpWebHost();
+            this.portGenerator = new Random();
+            this.attemptedPorts = new HashSet<int>();
         }
 
         public Uri InternalRpcUri { get; private set; }
@@ -59,10 +70,12 @@ namespace Microsoft.Azure.WebJobs.Extensions.DurableTask
             int numAttempts = 1;
             do
             {
-                int availablePort = this.GetAvailablePort();
+                int listeningPort = numAttempts == 1
+                    ? DefaultPort
+                    : this.GetRandomPort();
                 try
                 {
-                    this.InternalRpcUri = new Uri($"http://127.0.0.1:{availablePort}/durabletask/");
+                    this.InternalRpcUri = new Uri($"http://127.0.0.1:{listeningPort}/durabletask/");
                     var listenUri = new Uri(this.InternalRpcUri.GetLeftPart(UriPartial.Authority));
                     this.localWebHost = new WebHostBuilder()
                         .UseKestrel()
@@ -80,11 +93,9 @@ namespace Microsoft.Azure.WebJobs.Extensions.DurableTask
                         this.durableTaskOptions.HubName,
                         functionName: string.Empty,
                         instanceId: string.Empty,
-                        message: $"Failed to open local socket {availablePort}. This was attempt #{numAttempts} to open a local port.");
+                        message: $"Failed to open local port {listeningPort}. This was attempt #{numAttempts} to open a local port.");
+                    this.attemptedPorts.Add(listeningPort);
                     numAttempts++;
-                    var random = new Random();
-                    var millisecondsToWait = (int)Math.Round(random.NextDouble() * 1000);
-                    await Task.Delay(millisecondsToWait);
                 }
             }
             while (numAttempts <= maxAttempts);
@@ -99,6 +110,20 @@ namespace Microsoft.Azure.WebJobs.Extensions.DurableTask
 #endif
         }
 
+        private int GetRandomPort()
+        {
+            // Get a random port that has not already been attempted so we don't waste time trying
+            // to listen to a port we know is not free.
+            int randomPort;
+            do
+            {
+                randomPort = this.portGenerator.Next(MinPort, MaxPort);
+            }
+            while (this.attemptedPorts.Contains(randomPort));
+
+            return randomPort;
+        }
+
         public async Task StopAsync()
         {
 #if !FUNCTIONS_V1
@@ -108,29 +133,6 @@ namespace Microsoft.Azure.WebJobs.Extensions.DurableTask
             await Task.Yield();
 #endif
             this.IsListening = false;
-        }
-
-        private int GetAvailablePort()
-        {
-            // If we are able to successfully start a listener looking on the default port without
-            // an exception, we can use the default port. Otherwise, let the TcpListener class decide for us.
-            try
-            {
-                var listener = new TcpListener(IPAddress.Loopback, DefaultPort);
-                listener.Start();
-                listener.Stop();
-                return DefaultPort;
-            }
-            catch (SocketException)
-            {
-                // Following guidance of this stack overflow answer
-                // to find available port: https://stackoverflow.com/a/150974/9035640
-                var listener = new TcpListener(IPAddress.Loopback, 0);
-                listener.Start();
-                int availablePort = ((IPEndPoint)listener.LocalEndpoint).Port;
-                listener.Stop();
-                return availablePort;
-            }
         }
 
         private async Task HandleRequestAsync(HttpContext context)