+++ release_notes.md
@@ -1,5 +1,6 @@
 ## New Features
 
 ## Bug fixes
+- Fix issue with local RPC endpoint used by non-.NET languages on Windows apps (#1800)
 
 ## Breaking Changes